requirements.txt
Django==3.0.3
djangorestframework==3.11
histographer-analysis==0.2.0
gunicorn==20.0.4
requests
PyYAML
pandas